+++ app/helper/get_system_info.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+"""Returns information regarding the system the app is running on"""
+
+
+def get_system_info() -> None:
+    """Returns information regarding the system the app is running on"""
+    # TODO import nv-smi etc
+    # os.umask_result
+    return NotImplementedError