+++ app/model/infer_model.py
@@ -6,6 +6,28 @@
 # from torch import cuda, no_grad
 
 
+def infer_model(
+    # input: str, device: str
+) -> str:  # TODO
+    """Infers result from model given input"""
+
+    # TODO adopt to model and task
+    # if dataset == 'YAHOO':
+    # input = "Why is cheese so much better with wine?"
+    # return test_model(input, device)
+    # elif dataset == 'MRPC':
+    #   print(test_model('hallo', 'hedda'))
+
+    # infer_model(
+    #     input,
+    #     get_tokenizer(self.paramobj.model_full_name),
+    #     get_model(self.paramobj.model_full_name),
+    #     self["device"],
+    # )
+
+    return NotImplementedError
+
+
 def test_model(inputs: str, device: str) -> str:
     """
     Test model output before fine-tuning.
@@ -33,25 +55,3 @@ def test_model(inputs: str, device: str) -> str:
         return NotImplementedError
     else:
         return NotImplementedError  # "NO CUDA"
-
-
-def infer_model(
-    # input: str, device: str
-) -> str:  # TODO
-    """Infers result from model given input"""
-
-    # TODO adopt to model and task
-    # if dataset == 'YAHOO':
-    # input = "Why is cheese so much better with wine?"
-    # return test_model(input, device)
-    # elif dataset == 'MRPC':
-    #   print(test_model('hallo', 'hedda'))
-
-    # infer_model(
-    #     input,
-    #     get_tokenizer(self.paramobj.model_full_name),
-    #     get_model(self.paramobj.model_full_name),
-    #     self["device"],
-    # )
-
-    return NotImplementedError