+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]
 ---
 
+[1.9.0] - 2022-08-14
+---
+
 ### Added
 
 - `omegaconf` package to load YAML configuration files